+++ doc/changes.md
@@ -1,4 +1,5 @@
 # Change on author code
+## Dataset Inference Code Checking Exp
 - Change import in `models.py`
 ```python
 from model_src.preactresnet import *
@@ -26,3 +27,9 @@ change to `remaining[remaining.clone()] = new_remaining`
 - Add `utils/logger.py` script for print output
 - Add setup `Ubuffered` output stream code for `generate_features.py` and `attack.py`
 - Change `print` file direction for `generate_features.py` and `attack.py`
+
+## Use unrelated datasets
+- Add 'MINST' in `n_class` map
+- Add 'MINST' in `net_mapper` map for `train.py` & `generate_features.py`
+- Add new `--experiment` parser for quick setting
+- Add `dataset.py` for alternate dataloading